Change button "Activate" to an "more" action
