+++ contract_compassion/__openerp__.py
@@ -44,6 +44,7 @@ Compassion contract module.
         'view/contract_group_view.xml',
         'view/contract_origin_view.xml',
         'view/contract_view.xml',
+        'view/activate_contract_view.xml',
         'workflow/contract_workflow.xml',
         'security/ir.model.access.csv',
         'data/friday_invoicer_cron.xml',
