+++ contract_compassion/view/activate_contract_view.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <record id="view_activate_sponsorship" model="ir.ui.view">
+            <field name="name">contract.activate.form</field>
+            <field name="model">recurring.contract.activate.wizard</field>
+            <field eval="18" name="priority"/>
+            <field name="arch" type="xml">
+                <form string="Contract activation" version="7.0">
+                    <p>This will mark the contract as Active, even if the partner has made no payment. Please be sure that you want to do this.</p>
+                    <footer>
+                        <button name="activate_contract" string="Activate" type="object" class="oe_highlight"/>
+                        or
+                        <button string="Cancel" class="oe_link" special="cancel"/>
+                    </footer>
+                </form>
+            </field>
+        </record>
+
+        <!-- Menu Item -->
+        <act_window id="action_activate_contract"
+                key2="client_action_multi" name="Force activation"
+                res_model="recurring.contract.activate.wizard" src_model="recurring.contract"
+                view_mode="form" target="new" view_type="form"
+                />
+    </data>
+</openerp>
