+++ contract_compassion/wizard/__init__.py
@@ -10,3 +10,4 @@
 ##############################################################################
 
 from . import recurring_invoicer_wizard
+from . import contract_activation_wizard
